# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/workspaces/{ws}/databases/{db}/records/{rec}/comments": {
      "get": {
        "operationId": "CommentsController_list",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rec",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Comments, newest first",
        "tags": [
          "comments"
        ]
      },
      "post": {
        "operationId": "CommentsController_create",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rec",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentBodyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Comment (guests included); mentions extracted server-side",
        "tags": [
          "comments"
        ]
      }
    },
    "/api/v1/workspaces/{ws}/databases/{db}/records/{rec}/comments/{comment}": {
      "patch": {
        "operationId": "CommentsController_update",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rec",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "comment",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit own comment",
        "tags": [
          "comments"
        ]
      },
      "delete": {
        "operationId": "CommentsController_remove",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rec",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "comment",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete own comment (admins: any)",
        "tags": [
          "comments"
        ]
      }
    },
    "/api/v1/workspaces/{ws}/databases/{db}/records/{rec}/activity": {
      "get": {
        "operationId": "ActivityController_list",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rec",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "integer"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record activity trail, newest first (cursor)",
        "tags": [
          "activity"
        ]
      }
    },
    "/api/v1/workspaces/{ws}/databases/{db}/records/{rec}/attachments": {
      "get": {
        "operationId": "AttachmentsController_list",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rec",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Attachments on a record",
        "tags": [
          "attachments"
        ]
      },
      "post": {
        "operationId": "AttachmentsController_upload",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rec",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload a file (multipart field \"file\"; size-capped)",
        "tags": [
          "attachments"
        ]
      }
    },
    "/api/v1/workspaces/{ws}/databases/{db}/records/{rec}/attachments/{att}/download": {
      "get": {
        "operationId": "AttachmentsController_download",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rec",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "att",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Download the file (authz-checked, streamed)",
        "tags": [
          "attachments"
        ]
      }
    },
    "/api/v1/workspaces/{ws}/databases/{db}/records/{rec}/attachments/{att}/thumbnail": {
      "get": {
        "operationId": "AttachmentsController_thumbnail",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rec",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "att",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Image thumbnail (404 for non-images)",
        "tags": [
          "attachments"
        ]
      }
    },
    "/api/v1/workspaces/{ws}/databases/{db}/records/{rec}/attachments/{att}": {
      "delete": {
        "operationId": "AttachmentsController_remove",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rec",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "att",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an attachment (object removed best-effort)",
        "tags": [
          "attachments"
        ]
      }
    },
...
    "/api/v1/workspaces/{ws}/grants": {
      "get": {
        "operationId": "GrantsController_list",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List access grants (optionally for one user)",
        "tags": [
          "access"
        ]
      },
      "post": {
        "operationId": "GrantsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGrantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Grant a role on a space or database (upserts per scope)",
        "tags": [
          "access"
        ]
      }
    },
    "/api/v1/workspaces/{ws}/grants/{grant}": {
      "delete": {
        "operationId": "GrantsController_remove",
        "parameters": [
          {
            "name": "grant",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke a grant",
        "tags": [
          "access"
        ]
      }
    }
  },
  "info": {
    "title": "StoryOS API",
    "description": "The public StoryOS REST API. The web app is client #1 of this same API — everything the UI does, you can do.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CommentBodyDto": {
        "type": "object",
        "properties": {
          "body": {
            "minItems": 1,
            "maxItems": 200,
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "enum": [
                        "text"
                      ]
                    },
                    "text": {
                      "type": "string",
                      "maxLength": 5000
                    }
                  },
                  "required": [
                    "type",
                    "text"
                  ]
                },
                {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "enum": [
                        "mention"
                      ]
                    },
                    "user_id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "user_id"
                  ]
                }
              ]
            }
          }
        },
        "required": [
          "body"
        ]
      },
      "CreateGrantDto": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "minLength": 1
          },
          "space_id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "database_id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "role": {
            "type": "string",
            "enum": [
              "viewer",
              "commenter",
              "editor",
              "creator"
            ]
          }
        },
        "required": [
          "user_id",
          "role"
        ]
      }
    }
  }
}